+++ lib/search_app_bar_page.dart
@@ -4,6 +4,7 @@ export 'src/search_app_bar_page/ui/search_app_bar_page.dart';
 export 'src/search_app_bar_page/ui/search_app_bar_page_stream.dart';
 export 'src/search_app_bar_page/ui/search_app_bar_pagination.dart';
 export 'src/search_app_bar_page/controller/utils/filters/filters_type.dart';
+export 'src/search_app_bar_page/ui/infra/rx_get_type.dart';
 
 export 'src/search_app_bar_page/controller/simple_app_bar_controller.dart';
 export 'src/search_app_bar_page/ui/core/search_app_bar/search_app_bar.dart';